Big cleanup after clang-tidy run and code analysis.
There are some TODOs left which are marked with // TODO

Reenabled compilation flags for debug code.

Changed a lot of uint64 values to signed type. The signed type still
offers us the possibility to work with tremendously huge files and it
is easier in some operations to have a signed value like -1 for a file
size query (not existing, stream, whatever).

Removed unnecessary imports throughout the project.

Changed a lot of code so, that equal operations are correct. See to it:
- that compared types are both signed or unsigned
- compared types have the right sizes 32Bit < 64Bit and 64Bit<64Bit

Changed method parameter types to const ref where feasible.

Changed auto loop parameters to const ref where necessary.

Refactored Extractor::extract() and pulled out some methods to make it
more readable.

Renamed checkPremises throughout the whole project to fulfillsPremises.

Changed "old-style" casts to e.g. static_cast / reinterpret_cast over
all files.

Added more tests for the index entry storage strategies.

+++ src/process/io/IOBase.h
@@ -42,7 +42,7 @@ public:
 
     virtual bool exists() = 0;
 
-    virtual u_int64_t size() = 0;
+    virtual int64_t size() = 0;
 
     virtual bool empty() = 0;
 
@@ -55,16 +55,16 @@ public:
      * @param bytes
      * @return Number of rewound bytes
      */
-    virtual int seek(int64_t nByte, bool absolute) = 0;
+    virtual int64_t seek(int64_t nByte, bool absolute) = 0;
 
     /**
      * Seek forward the amount of nByte from the current position.
      * @param nByte Number of Bytes to skip forward.
      * @return The number of Bytes which were skipped.
      */
-    virtual int skip(uint64_t nByte) = 0;
+    virtual int64_t skip(int64_t nByte) = 0;
 
-    virtual int rewind(uint64_t nByte) {
+    virtual int64_t rewind(int64_t nByte) {
         return seek(-nByte, false);
     }
 
@@ -74,7 +74,7 @@ public:
      * Get the current position in the source.
      * @return
      */
-    virtual uint64_t tell() = 0;
+    virtual int64_t tell() = 0;
 
     virtual int lastError() = 0;
 };
